feat: hide SC, small site fixes, remove SC tests by default [MDS-483]
