+++ next-docs/components/sidebar/Sidebar.tsx
@@ -11,7 +11,7 @@ type Props = {
 };
 
 const Sidebar: React.FC<Props> = ({ closeSidebar }) => (
-  <div className="z-10 fixed top-0 h-screen w-80 flex flex-col flex-grow gap-6 pt-12 pb-[72px] px-5 lg:px-8 bg-goku overflow-y-scroll">
+  <div className="z-10 fixed top-0 h-screen w-80 flex flex-col flex-grow gap-6 pt-12 pb-24 px-5 lg:px-8 bg-goku overflow-y-scroll">
     <div className="flex items-center flex-shrink-0 ps-3 text-bulma">
       <LogoLink />
     </div>
