+++ next-docs/components/sidebar/Version.tsx
@@ -2,7 +2,7 @@ import { LogoMoonDesignShort } from '@heathmont/moon-assets';
 import packageInfo from '../../package.json';
 
 const Version = () => (
-  <div className="fixed bottom-0 left-0 w-80 py-6 px-8 flex text-moon-14 bg-goku gap-2">
+  <div className="fixed bottom-0 left-0 w-80 py-6 px-8 flex text-moon-14 bg-goku gap-2 border-t border-t-beerus">
     <LogoMoonDesignShort fontSize="1.5rem" />
     <p>
       Version <span className="font-medium">{packageInfo.version}</span>
