+++ next-docs/package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start -p 80",
-    "test": "jest --watch",
+    "test": "jest --watch --testPathPattern=public/examples",
+    "test:styled": "jest --watch --testPathPattern=public/styledExamples",
     "test:update": "jest -u",
     "test:ci": "jest -w 1 --ci",
     "build:template:css": "npx tailwindcss -i ./styles/globals.css -o ./styles/output.css",
